Revert "Reorganize components into folders"
