+++ src/App.tsx
@@ -3,18 +3,18 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import "./App.css";
 import { Amplify } from "aws-amplify";
 import awsconfig from "./aws-exports";
-import Success from "./components/authentication/success";
-import ResetPassword from "./components/authentication/resetPassword";
-import CreateAccount from "./components/authentication/createAccount";
-import EnterCode from "./components/authentication/enterCode";
-import Login from "./components/authentication/login";
-import ForgotPassword from "./components/authentication/forgotPassword";
-import TimeSlot from "./components/popup/timeslot";
+import Success from "./components/success";
+import ResetPassword from "./components/resetPassword";
+import CreateAccount from "./components/createAccount";
+import EnterCode from "./components/enterCode";
+import Login from "./components/login";
+import ForgotPassword from "./components/forgotPassword";
+import TimeSlot from "./components/timeslot";
 import Calendar from "./components/calendar";
-import CalendarMobile from "./components/mobile/mobileCalendar";
-import MobileTimeslots from "./components/mobile/mobileTimeslots";
-import TimeslotSuccess from "./components/popup/timeslotSuccess";
-import TimeSlotConfirmation from "./components/popup/timeslotConfirmation";
+import CalendarMobile from "./components/mobileCalendar";
+import MobileTimeslots from "./components/mobileTimeslots";
+import TimeslotSuccess from "./components/timeslotSuccess";
+import TimeSlotConfirmation from "./components/timeslotConfirmation";
 
 Amplify.configure(awsconfig);
