+++ src/components/resetPassword.tsx
@@ -2,9 +2,9 @@
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { Auth } from "aws-amplify";
-import eyeSlash from "../../images/eyeSlash.svg";
-import eye from "../../images/eye.svg";
-import backArrow from "../../images/backArrow.png";
+import eyeSlash from "../images/eyeSlash.svg";
+import eye from "../images/eye.svg";
+import backArrow from "../images/backArrow.png";
 import {
   Wrapper,
   Box,
@@ -17,7 +17,7 @@ import {
   PasswordContainer,
   EyeSlash,
   ErrorMessage,
-} from "../styledComponents";
+} from "./styledComponents";
 
 // email prop that is set in forgot password page
 type EmailProps = {
